+++ src/Models/WalletModel.h
@@ -133,7 +133,7 @@ signals:
 
 
 public slots:
-    Q_INVOKABLE bool transfer(double amount, const QString& address, int pFee, const QString& pPaymentId);
+    Q_INVOKABLE bool transfer(double amount, const QString& address, double pFee, const QString& pPaymentId);
